+++ content/themes/air-blocks/inc/hooks/gutenberg.php
@@ -5,7 +5,7 @@
  * @Author: Niku Hietanen
  * @Date: 2020-02-20 13:46:50
  * @Last Modified by:   Roni Laukkarinen
- * @Last Modified time: 2021-06-22 10:57:39
+ * @Last Modified time: 2022-02-08 18:25:00
  *
  * @package airblocks
  */
@@ -85,10 +85,14 @@ function register_block_editor_assets() {
 // Remove Gutenberg inline "Normalization styles" like .editor-styles-wrapper h1
 // color: inherit;
 // @source https://github.com/WordPress/gutenberg/issues/18595#issuecomment-599588153
-function remove_gutenberg_inline_styles( $editor_settings, $post ) {
-  unset( $editor_settings['styles'][0] );
+// @ref https://gist.github.com/gziolo/a947dc52eb2604c77a0a5b0797b2e781#block_editor_settings_all
+function remove_gutenberg_inline_styles( $editor_settings, $editor_context ) {
+  if ( ! empty( $editor_context->post ) ) {
+    unset( $editor_settings['styles'][0]['css'] );
+  }
+
   return $editor_settings;
-} // end remove_gutenberg_inline_styles
+}
 
 /**
  * Make sure Gutenberg wp-admin editor styles are loaded
@@ -119,79 +123,43 @@ function block_editor_title_input_styles() {
   .edit-post-visual-editor__post-title-wrapper + .is-root-container > .wp-block:first-child {
     margin-top: 0;
   }
-
   /* Remove white border from top */
   .interface-interface-skeleton__header {
     border-bottom: 0;
   }
-
   .block-editor .editor-styles-wrapper {
     padding-top: 0;
   }
-
   .block-editor .editor-styles-wrapper .edit-post-visual-editor__post-title-wrapper {
     background-color: #23282e;
     border-bottom: 1px solid #23282e;
+    color: #fff;
     position: relative;
     z-index: 3;
-    color: #fff;
   }
-
-  .block-editor .editor-styles-wrapper .edit-post-visual-editor__post-title-wrapper .components-visually-hidden::after {
-    /* content: "(näkyy mm. valikossa, selainikkunan nimessä ja murupolussa)"; */
-    content: "(is shown for example in navigation, browser window name and in breadcrumbs)";
-    color: rgba(255, 255, 255, .5);
-    display: inline;
-    margin-left: 5px;
+  .edit-post-visual-editor__post-title-wrapper {
+    margin: 0;
   }
-
   .block-editor .editor-styles-wrapper .editor-post-title {
-    padding: 4rem 2rem;
-    margin: 0 auto;
-  }
-
-  .block-editor .editor-styles-wrapper .editor-post-title .components-visually-hidden {
-    border: initial;
-    clip: initial;
-    -webkit-clip-path: initial;
-    clip-path: initial;
-    color: rgba(255, 255, 255, .5);
-    font-size: var(--font-size-15);
-    display: block;
-    height: initial;
-    margin: initial;
-    margin-bottom: 1.2rem;
-    overflow: initial;
-    padding: initial;
-    position: initial;
-    width: initial;
-    word-wrap: initial;
-  }
-
-  .block-editor .editor-styles-wrapper .editor-post-title .editor-post-title__input {
-    line-height: 1.3;
-    background-color: #2e3338;
-    border-color: #2e3338;
-    color: rgba(255, 255, 255, .5);
-    border-radius: 3px;
-    border-style: solid;
-    border-width: 2px;
-    box-sizing: border-box;
-    font-family: inherit;
-    font-size: var(--font-size-22);
-    font-weight: 400;
-    line-height: 24px;
-    margin-bottom: 0;
-    padding: 15px;
-    position: relative;
-    text-decoration: none;
-    transition: all 0.55s;
-  }
-
-  .block-editor .editor-styles-wrapper .editor-post-title .editor-post-title__input:focus {
-    background-color: #000;
-    border-color: #000;
     color: #fff;
+    font-size: 20px;
+    font-weight: 600;
+    margin: 0 auto;
+    padding: 4rem 2rem;
+  }
+  .block-editor .editor-styles-wrapper .editor-post-title::before {
+    color: rgb(255 255 255 / .5);
+    display: block;
+    font-size: 12px;
+    font-weight: 500;
+    margin-bottom: 1rem;
+    position: relative;
+  }
+  body.locale-fi .editor-styles-wrapper .editor-post-title::before {
+    content: "Nimi, joka näkyy selaimen välilehdessä ja valikossa";
+  }
+  body.locale-en-us .editor-styles-wrapper .editor-post-title::before {
+    content: "Post name shown in the browser tab and menus";
   }
   ';
   wp_add_inline_style( 'block-editor-styles',  $styles );