+++ bin/newblock.sh
@@ -3,11 +3,11 @@
 # @Author: Roni Laukkarinen
 # @Date:   2021-11-23 18:06:59
 # @Last Modified by:   Roni Laukkarinen
-# @Last Modified time: 2022-02-08 16:46:12
+# @Last Modified time: 2022-02-10 16:03:39
 
 # Script specific vars
 SCRIPT_LABEL='for macOS'
-SCRIPT_VERSION='1.1.0'
+SCRIPT_VERSION='1.1.1'
 
 # Vars needed for this file to function globally
 CURRENTFILE=`basename $0`