+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### [Unreleased]
+### 1.0.4: 2021-02-10
 
 * Fix 5.9 related WSOD
 * Disable air-helper media directory hook as it breaks things in production